config.ts: Fix getURL not merging site-specific config objects
https://github.com/tridactyl/tridactyl/issues/1193 happened because
before this commit, getURL did not merge objects from different
site-specific configs. This means that if a user had the following site
specific configs:

"example": {
        "priority": 1,
        "nmaps": {
                "a": "js alert('')"
        },
}
"example.org": {
        "priority": 2,
        "nmaps": {}
}

A call to config.getURL("example.org", ["nmaps"]) resulted in the empty
object from the "example.org" config being returned and the mappings for
"example" to be ignored. This obviously is not the desired behavior.

This is fixed by mergeDeep()'ing objects found in matching site-specific
configs.

Closes #1193.