lib/editor.ts: Make editor functions work on input where type = "email"
Before this commit, editor functions that read or modified caret
position in input fields did not work on input fields the type of which
was "email". This was because browsers implement the selectionStart and
selectionEnd attributes only for a few input types.
This problem is fixed by first converting the input field to text,
applying the wanted modifications and then changing the input field's
type back to whatever it was before.

Closes #1209.