hinting.ts: Move overlapping hints around
This commit makes Tridactyl move overlapping hints around in order to
try to untangle them. This is pretty much a failure, hints are still
overlapping and when they aren't it's hard to tell what hint belongs to
what element. One possibility is that the magic numbers in this commit
(`repulsionForce`, `attraction` and `velocityLoss`) just need to be
tweaked until a sweet spot is found, but it is more probable that moving
hints around is just a bad idea.

This code might still be useful for an april's fools if we set
`velocityLoss` to 1 though.