commandline_frame.ts: stop using dirty hacks for completions
The previous code simulated an input event in order to trigger the input
event handler which recomputed completions. This was ok until delays
were added to the input event handlers in order to reduce the lag that
could happen when typing fast/keeping a key pressed. This delay also
affects completion computation on other actions, such as fillcmdline.

In order to remove this delay, we move completion computation out of the
event handler and directly call this functions everywhere we previously
triggered an input event.

This should help with https://github.com/tridactyl/tridactyl/issues/1242