commandline_frame.ts: update HISTORY_SEARCH_STRING when needed
https://github.com/tridactyl/tridactyl/issues/1329 describes a bug that
can be triggered with the following key presses:

:t<ArrowUp><ArrowDown><Backspace>w<ArrowUp>

This results in the commandline being filled with a command starting
with `t` rather than `w`. This is caused by not updating the
HISTORY_SEARCH_STRING variable on changes and is fixed by always
resetting it if the previous command did not call history().

Closes #1329.