Resize command line only once per available completion source
One of the problems of the command line was that it made a resizeArea()
call for each enabled completion, no matter whether its status was
"hidden" or "normal". This was a problem because a resizeArea call
results in 2 cross-script messages: a "show" and a "focus" message. This
means that for each keystroke, we sent 28 messages. This commit fixes
that thanks to modifications in multiple files:

- commandline_frame.ts: Stop accumulating event listeners on resizeArea
  calls. Make sure completion sources actually need a refresh before
  calling resizeArea().
- completions.ts: Add logic to know whether a completion source needs a
  refresh or not.
- {Rss,Sessions,Tab,TabAll,Window}.ts: Make sure that completions are
  actually needed before computing them.

This seems to make opening the command line slightly faster for me,
although I can't tell if this is placebo or not.