wine-pyinstaller.sh: Fix checkPrerequisite()
The problem was the `[ ! -f "${bin_loc}" ]` check. `command -v` does not
necessarily return an absolute path (it can just return the name of the
command for builtins) and this is what happened for 'printf'. Thus, the
`[ ! -f` check failed.
This check doesn't really make sense. If `command -v $binary` doesn't
fail you're good to go, it doesn't really matter where $binary is
located.