Remove BGSELF
BGSELF was a hack that I used when implementing ex commands for the
command line. It consisted of having .excmds_background.generated.ts
import itself as BGSELF in order to dynamically add commands to it. This
let us define excmds in other files while not changing anything in
parsers/exmode.ts.

This was awful so I decided to remove it. This required performing the
following changes:
- Moving text.* and ex.* command definitions to their own files where
  they have zero side effects. While this was easy for text commands, ex
  commands rely a lot on side effects. In order to work around this,
  lib/commandline_cmds exports a single function, getCommandlineFns(),
  which expects an object representing the commandline's state as
  parameter.
- In the background script, import our side effect free files and wrap
  them in proxys that will send "commandline_cmd" and "editorfn_content"
  messages to tabs when needed.
- In the content script, add a listener that will either execute an
  editor function or forward it to the command line when it receives an
  "editorfn_content" message.
- In the commandline script, add a listener that will execute exmode
  commands.