Refactor DownloadPost to send pertinent data.
This addresses issues #810 and #807. An optional argument of the type
AutocmdArgs can be passed to loadaucmds. Only a small subset of the variables
mentioned in #810 are implemented. I think it's better to implement these as the
need/request arrives but for the DownloadPost aucmd I felt like this was the
most straightforward way to implement.

The parseaucmdargs function takes an aucmd and looks for the pattern "<param">
where param is any of the keys provided by AutocmdArgs and replaces the pattern
with the matching information from the args.

Example usage:
`:au DownloadPost .* fillcmdline "<file>"`
Will print the full path of the downloaded file on the cmdline.