Start factoring excmd implementations out into libraries
We use glacambre's trick with the Proxy and the `infer U` to permit
type inference to propagate through the wrapper that forwards commands
to the right context. This _should_ allow us to eventually replace the
functionality in excmd_macros to mangle separate background and
content excmd files from the global exmcds.ts file, as that forwarding
will be done on the way to the libraries rather than at the top level.