excmds.test.ts: Add `:guiset` test
This test makes sure guiset has completions and actually writes
something in the user's userChrome.
Writing this test made me discover that `-profile` was a valid firefox
argument and so I added support for that in native.ts. This made me
realize that the profile detection code didn't handle profiles that
aren't described in `profile.ini` and so I fixed that.