scrolling.ts: fix smooth scrolling
This commit fixes smooth scrolling. The issue with smooth scrolling was
that from time to time, pressing `j` would only scroll the page by a
pixel. This happened because the computed scrolling step was smaller
than a pixel. When writing to elem.scrollTop, this value would be
rounded to the nearest integer, which sometimes was equal to the
previous scrollTop value. This resulted in Tridactyl believing that the
element couldn't be further scrolled and stopping there.

This is fixed by using Math.ceil()/round() on the step in order to make
sure a round value is obtained. This value is then compared to the
current scrollTop value and incremented in case they're equal, in order
to make sure scrolling makes progress.

Fixes https://github.com/tridactyl/tridactyl/issues/627 and maybe
https://github.com/tridactyl/tridactyl/issues/54 ?