Implement writerc native messenger call
This allows us to write the rc to OS. The only way to overwrite your current
rc file is to use mkt!, mktridactylrc! or mktridactylrc -f. mkt* commands
issued without the force/exclamation mark will not overwrite existing files.

Native messenger was updated to 0.1.11 and the mktridactylrc command is
unavailable in prior versions.

Fixed some random indentation errors that occurred in earlier commit.