Add shortcut for ignore mode
When I started using tridactyl, I didn't find any shortcut to the ignore mode that was viable for laptop keyboards (at least on mine, since I haven't got backticks and `Ins` must be activated with `Fn`). `Shift-Esc` is relatively harmless and pretty easy to remember, so I think it may be a good one for beginners.