Fix storageloc setting
The storageloc setting is inherently unsyncable; since it is used to
decide which storage to use, it is always set to "local" in local
storage and "sync" in sync storage. Any updates from sync storage will
cause mayhem and override local settings with "sync".

So instead, ensure that it is always saved to local storage with the
intended value, and that remote updates cannot change it.