Fix and optimize addJump scroll handler
It was reading properties that a scroll event does not have.

It was messaging the background page with an expensive call on every
scroll event. This slowed down the page unnecessarily.

Also extracted jumps.timeoutid to the content script scope.
This variable is not useful to others outside of the current content
script. Storing it locally means we can avoid expensive communication
with the background page and only do it once we're sure we need to add
a new entry to the jump list.