getBestEditor(): Extract os constant, define arg_quote and vim_positioning_arg
In order to apply the cursor positioning for the Vim editor as
documented for 'editorcmd', define a constant to avoid duplicating the
(rather complex) argument for all of the mvim / nvim / vim / gvim
variants. The argument needs to be quoted (at least on Unix, and there
single quotes have to be used to avoid history expansion (e.g. in Bash)
triggered by the "!" character), so another platform-specific constant
(arg_quote) is needed (as Windows only supports double quotes).