Fix type error when calling contextualIdentities.update
The error is:

```
controller in excmd: Error: Type error for parameter
details (Unexpected property "cookieStoreId") for
contextualIdentities.update.
```

This is because container.update invokes `fromString` on the user's
input, which produces an object with `cookieStoreId` set. If you look
at `fromString` you'll note that it has to some dirty type coercion to
make its return type work out.