Remove Windows Terminal from default Windows terminals again
Turns out WT, when run, apparently spawns another instance of itself and then immediately exits, with the newly spawned instance being the one that's actually displayed. Or something like that. That makes it not suited as a terminal for `:editor` after all, because once the first process exits, Tridactyl will think the user is finished writing and read the file, resulting in an empty text box. There doesn't seem to exist any option to pass to wt.exe to make it not do that, so we can't use it.