Native, Windows: fix 'nativeinstallcmd' for Windows 8 and earlier
To be able to download the installer, Powershell must be able to use a
security protocol that GitHub supports. Before Windows 10, we have to
set it.  We do the same thing in the installer already, but we have to
do it in both places so the installer works both when invoked as part
of 'nativeinstallcmd' and and on its own.