Fix #4158: bookmarklet execution bug
In original implementation, a script tag will leave in html,
and the bookmarklet in URL Encoding will not work.
Moreover, if bookmarklet return a string after evaluation,
the string should replace current page's html content.

One can execute bookmarklet by assign it to `window.location.href`,
but it does not work in add-on's context,
so I assign it with `window.eval()`.