Add --url option to define command in certain url and fix a bug
Accept url option in command and output them in mktridactylrc
with --url option. The doc about it and a example is added too.

The original code does not detect the recursive alias correctly.
Thought it does detect, it detect before it define the alias;
it should detect it after defining.