Fix setting new config value in help page not working bug
In `moz-extension://*/static/docs/classes/_src_lib_config_.default_config.html`,
set a new value in `<input>` and press Enter should set a new value,
but it does not work, because the sent message payload has a wrong type,
which should be a string instead of a array.

The *Reset Tridactyl config* button is also fixed.