Fix mergeDeep with Object.assign pollute the object type values
The original mergeDeep call the Object.assign on `r[key]`
if r[key] is a object, but if so, r[key] and o2[key] will be the
same object, and assign values to it will modify the o2[key].

In my situation, the o2 is a subconfig, and it is polluted with the
global config (o1) when merging.
I can not exactly test or locate the problem introduced by this fix/bug.