Fix wrong scrolling when the match not in root scroll element
If the match text is inside a scrollable element which is not root,
the highlight rectangle will be in a wrong position, because
the position is relative the root element but not the scrolling element,
and scroll to the hightlight rectangle will not make the matched text visible.

This commit make focus method scroll to the match text range instead of the
highlight rectangle by a scrollIntoView polyfill library, which does not
check the input strictly, and the highlight rectangle position will update
after the scroll is performed.