Show '#' only if last active tab was in this group
Alternate tab (for `:tab #`) is always the last active tab, even
if it's hidden, not the last active visible tab. This means the
alternate tab is not listed if it's in a different group and
`tabshowhidden==true`.

(It's still possible to switch to a tab that is not displayed in
completions.)