Try to fix more typing errors
The messages object in background.ts implicitly had a type
{ [key: string]: { [key: string]: (...args: any) => any } },
and was used as such in messaging.ts. Now that TypeScript is updated, it
got angry because the imported excmds_background and statsLogger are not
strictly { [key: string]: (...args: any) => any }. For excmd_functions
the problem were consts which are not functions. I namespaced all
functions and if we import only that namespace and use it in messaging,
typescript will be happy. I know namespaces are discouraged but i don't know
of other ways to separate functions from other parts of a file without
creating a new file for them. As for logger, i used a nasty cast because
what else can we do?
Using namespace also pisses off lint, and pretty seems to break
everything, so this patch is purely to make build work.