allow focusinput focus node in shadow-root
It seem that document.contains is to check whether
the node is removed, but not to consider about the
shadow root.

The isConnected property mean whether the node is
connect to a document or a shadow root that
connected to the root document.
It is enough to check whether the document is
removed.